// lib/types.ts
export interface UserJwtPayload {
    userId: string;
}

export interface UserRecord {
    email: string;
    name: string;
    phone?: string;
    address?: string;
    usertype: number;
    shopid?: ShopBrief[];
    password?: string;
    tobeowner: boolean;
    created: string;
    tobeownerAt?: string;
}
// ...
export interface NoAuthorization{
    no_authorization: number;
}
// ...
export interface ProcessResult{
    status: string;
    message: string;
}

export interface MailOptions {
    from: string;
    to: string;
    subject: string;
    html: string;
}

export interface ImgDimensionType{
    width: number;
    height: number;
}

export interface ShopDataRecord{
    shopname: string;
    foodsupply: string;
    profileimage: string;
    owner: string; 
    onboard: boolean;
    created?: string;
} 

export interface ShopDataType{
    id: string;
    shopname: string;
    foodsupply: string;
    profileimage: string;
    owner: string; 
    onboard: boolean;
    blockusers?: string[];
    created?: string;
} 

export interface ShopBrief{
    id: string;
    shopname: string;
    onboard: boolean;
}